cdk/requirements.txt
appdirs==1.4.4
attrs==20.3.0
aws-cdk.assets==1.83.0
aws-cdk.aws-apigatewayv2==1.83.0
aws-cdk.aws-applicationautoscaling==1.83.0
aws-cdk.aws-autoscaling-common==1.83.0
aws-cdk.aws-certificatemanager==1.83.0
aws-cdk.aws-cloudformation==1.83.0
aws-cdk.aws-cloudwatch==1.83.0
aws-cdk.aws-codeguruprofiler==1.83.0
aws-cdk.aws-dynamodb==1.83.0
aws-cdk.aws-ec2==1.83.0
aws-cdk.aws-ecr==1.83.0
aws-cdk.aws-ecr-assets==1.83.0
aws-cdk.aws-efs==1.83.0
aws-cdk.aws-events==1.83.0
aws-cdk.aws-iam==1.83.0
aws-cdk.aws-kms==1.83.0
aws-cdk.aws-lambda==1.83.0
aws-cdk.aws-lambda-python==1.83.0
aws-cdk.aws-logs==1.83.0
aws-cdk.aws-route53==1.83.0
aws-cdk.aws-s3==1.83.0
aws-cdk.aws-s3-assets==1.83.0
aws-cdk.aws-sns==1.83.0
aws-cdk.aws-sqs==1.83.0
aws-cdk.aws-ssm==1.83.0
aws-cdk.cloud-assembly-schema==1.83.0
aws-cdk.core==1.83.0
aws-cdk.custom-resources==1.83.0
aws-cdk.cx-api==1.83.0
aws-cdk.region-info==1.83.0
black==20.8b1
cattrs==1.1.2
click==7.1.2
constructs==3.2.99
jsii==1.16.0
mypy-extensions==0.4.3
pathspec==0.8.1
publication==0.0.3
python-dateutil==2.8.1
regex==2020.11.13
six==1.15.0
toml==0.10.2
typed-ast==1.4.2
typing-extensions==3.7.4.3
